Integer price replaced by BigDecimal

+++ market-cart/api/src/main/java/com/github/mdpetrenko/market/cart/dto/CartDto.java
@@ -1,16 +1,17 @@
 package com.github.mdpetrenko.market.cart.dto;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 
 public class CartDto {
     private Collection<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
     private String id;
 
     public CartDto() {
     }
 
-    public CartDto(Collection<CartItemDto> items, int totalPrice, String id) {
+    public CartDto(Collection<CartItemDto> items, BigDecimal totalPrice, String id) {
         this.items = items;
         this.totalPrice = totalPrice;
         this.id = id;
@@ -24,11 +25,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
